pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = [
    "Authors Here"
]

[tool.poetry.dependencies]
python = "^3.7"
flask_jwt_extended = "^3.24.1"
bcrypt = "^3.1.7"
flask_pymongo = "^2.3.0"
flask_cors = "^3.0.8"
flask_restx = "^0.2.0"
flask_script = "^2.0.6"
python-decouple = "^3.3"
python-dotenv = "^0.13.0"
gunicorn = "^20.0.4"
pytest = "^5.4.2"
flask_testing = "^0.8.0"
flask_mail = "^0.9.1"
dnspython = "^1.16.0"
celery = "^4.4.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.2"
autopep8 = "^1.5.2"

[tool.poetry.scripts]
app = 'manage:runserver'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

